// src/scale.d.ts
export declare namespace ScaleType {
    const LINEAR: 'linear';
    const LOG: 'log';
    const POW: 'pow';
    const SQRT: 'sqrt';
    const TIME: 'time';
    const UTC: 'utc';
    const SEQUENTIAL: 'sequential';
    const QUANTILE: 'quantile';
    const QUANTIZE: 'quantize';
    const THRESHOLD: 'threshold';
    const ORDINAL: 'ordinal';
    const POINT: 'point';
    const BAND: 'band';
}
export declare type ScaleType = typeof ScaleType.LINEAR | typeof ScaleType.LOG | typeof ScaleType.POW | typeof ScaleType.SQRT | typeof ScaleType.TIME | typeof ScaleType.UTC | typeof ScaleType.SEQUENTIAL | typeof ScaleType.ORDINAL | typeof ScaleType.POINT | typeof ScaleType.BAND;
// ...
export declare const defaultScaleConfig: {
    round: boolean;
    textXRangeStep: number;
    rangeStep: number;
    pointPadding: number;
    bandPaddingInner: number;
    facetSpacing: number;
    useRawDomain: boolean;
};